Change 1 – Added the association between the Repository and the catalog and dictionary:
Change 2 – An additional “Code” metaclass within DataDictionary has been removed.

Change 3 – Fix to the DictionaryItem in the eCore model to make it abstract.

Change 4 – RepositoryItem has been given a name attribute.

 

Change 5 – Added slots to DictionaryItem and CatalogItem in the eCore model, which were previously missing.

Change 6 – Message has been removed. This is because Jean-Marie pointed out this is intended to be a M0 concept.

Change 7 – DataType added as a kind of DictionaryItem. This is the only way DataTypes can be added to the Dictionary.

Change 8 – BusinessAssociation in the eCore model was corrected to match the MOF model. Previously it pointed at other BusinessAssociations instead of BusinessComponents.