Updated to include the dyanmic model proposal of MDR on the same day.
Dear WG4 – Based on the comments to date, please find my comments on the business process metamodel. These comments are for our next discussion, we won’t have time this week.

 

One of our first requests was to make a distinction between BusinessActor and BusinessRole. The v1 Standard is not clear in this regard, and so the RA implemented it as “Business Actor/Role”. This came up in the Catalogue of Differences, where the understandable response was that this needs defining. The definition I offer is that BusinessRole remains the type of the Participants behaviour, whereas Actor identifies a specific Business Role in a Business Transaction. For example when Broker A quotes a securities price to Broker B, both A and B has the BusinessRole of Broker, but Broker A is BusinessActor QuoteMaker and Broker B is BusinessActor QuoteTaker. This is necessary to distinguish different participants in a conversation:BusinessTransaction that have the same BusinessRole. Another example is - all WG4 members play the BusinessRole of ExpertMember, but there is only one BusinessActor of Secretary. 
The metaclass Interaction was added a few months ago. This was previously implicit and is now recorded explicitly. An Interaction has only one Sender, but has zero or more Receivers (in line with Part 6). MessageDefinition replaces Message as the type of the message in the Interaction. This was agreed the last two week and is based on Jean-Marie’s feedback on Message being an M0 concept. 
BusinessTransactions are composable. This was previously unclear in the Standard. This is necessary to support nested and iterative events – such as matching, quoting, and any other sort of negotiation. 
Should MessageDefinition be contained within a BusinessArea, a Business Process, or the entire BusinessProcessCatalog? 
Should we add the correlation from the Interaction to the conversation:BusinessTransaction? How do we match a message back to its conversation? Is this a v1.5 or v2.0 feature? 
The subConversation:BusinessTransactions currently contain no ability to represent loops, choices, parallel activities, etc. Should those be added in v1.5?