Issue #160
Background: Currently an instance of BusinessSubType is a 
TopLevelDictionaryEntry, and thus is included in the DataDictionary by 
being directly owned (via composition) by the DataDictionary.  It would be 
preferable for an instance of BusinessSubType to be owned by the 
BusinessComponent that is pointed to by the BusinessSubType's "subType" 
Property. Ownership by the BusinessComponent pointed to by the "subType" 
Property has the advantage of mandating that, if that BusinessComponent is 
deleted from the Repository, then that instance of BusinessSubType must be 
deleted as well.  Since BusinessSubType would no longer be a kind of 
TopLevelDictionaryEntry, BusinessSubType would be a subclass of 
RepositoryConcept, not of TopLevelDictionaryEntry.  An instance of 
BusinessSubType would be included in the DataDictionary by virtue of its 
ownership by a BusinessComponent, which is a TopLevelDictionaryEntry.

Recommendations: 

1) Change the AggregationKind of BusinessSubType's "subType" property to 
COMPOSITE

2) Make BusinessSubType a subclass of RepositoryConcept, not of 
TopLevelDictionaryEntry